requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2025.10.5
    # via
    #   pyogrio
    #   pyproj
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dateparser==1.2.2
    # via -r requirements.in
fonttools==4.60.1
    # via matplotlib
geopandas==1.1.1
    # via -r requirements.in
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.7
    # via -r requirements.in
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   geopandas
    #   matplotlib
    #   pandas
    #   pyogrio
    #   shapely
packaging==25.0
    # via
    #   geopandas
    #   matplotlib
    #   pyogrio
pandas==2.3.3
    # via
    #   -r requirements.in
    #   geopandas
pillow==12.0.0
    # via matplotlib
pyogrio==0.11.1
    # via geopandas
pyparsing==3.2.5
    # via matplotlib
pyproj==3.7.1
    # via geopandas
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   matplotlib
    #   pandas
pytz==2025.2
    # via
    #   dateparser
    #   pandas
pyyaml==6.0.3
    # via -r requirements.in
regex==2025.11.3
    # via dateparser
ruamel-yaml==0.18.15
    # via -r requirements.in
ruamel-yaml-clib==0.2.14
    # via ruamel-yaml
shapely==2.1.2
    # via geopandas
six==1.17.0
    # via python-dateutil
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via dateparser
utm==0.8.1
    # via -r requirements.in
